// src/app/posty-birb/models/gallery-status.model.ts
export enum GalleryStatus {
  NEW = 'NEW', //WHERE THERE IS A NEW OF THE ITEM
  COPY = 'COPY', //WHEN COPYING DATA ONLY
  UPDATE = 'UPDATE', //WHEN THE DATA TO A ITEM HAS BEEN UPDATED
  DELETE = 'DELETE', //WHEN AN ITEM HAS BEEN OR IS BEING DELETED
  EDITING = 'EDITING', //WHEN AN ITEM IS EMITTED FOR EDITING
  SCHEDULE = 'SCHEDULE CHANGE',
  QUEUED = 'QUEUED', //WHEN AN ITEM IS QUEUED FOR POSTING
  CANCEL = 'CANCEL' //WHEN THE POSTING FOR AN ITEM HAS BEEN CANCELED
}